[refactor/#1] 코드리뷰 기반 리팩토링
