+++ app/src/main/AndroidManifest.xml
@@ -13,28 +13,23 @@
         android:theme="@style/Theme.ANDANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".MyActivity"
-            android:exported="true"
-            android:label="@string/title_activity_my"
-            android:theme="@style/Theme.ANDANDROID" />
+            android:name=".presentation.ui.auth.MyActivity"
+            android:exported="false"
+            android:label="@string/title_activity_my" />
         <activity
-            android:name=".SignInAcitivity"
-            android:exported="true"
-            android:label="@string/title_activity_sign_in_acitivity"
-            android:theme="@style/Theme.ANDANDROID" />
+            android:name=".presentation.ui.auth.SignInAcitivity"
+            android:exported="false"
+            android:label="@string/title_activity_sign_in_acitivity" />
         <activity
-            android:name=".SignUpActivity"
-            android:exported="true"
-            android:label="@string/title_activity_sign_up"
-            android:theme="@style/Theme.ANDANDROID" />
+            android:name=".presentation.ui.auth.SignUpActivity"
+            android:exported="false"
+            android:label="@string/title_activity_sign_up" />
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.ANDANDROID">
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
