+++ app/src/main/java/org/sopt/and/presentation/ui/main/MainActivity.kt
@@ -1,0 +1,25 @@
+package org.sopt.and.presentation.ui.main
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import org.sopt.and.presentation.ui.auth.SignUpActivity
+import org.sopt.and.ui.theme.ANDANDROIDTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            ANDANDROIDTheme {
+                startActivity(
+                    Intent(this, SignUpActivity::class.java).apply {
+                        flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
+                    }
+                )
+            }
+        }
+    }
+}
