+++ app/src/main/java/org/sopt/and/presentation/utils/Constants.kt
@@ -1,0 +1,13 @@
+package org.sopt.and.presentation.utils
+
+object Constants {
+    //이메일, 비밀번호
+    const val KEY_EMAIL = "email"
+    const val KEY_PASSWORD = "password"
+
+    //회원가입 이메일, 비밀번호 길이 조건
+    const val MIN_PASSWORD_LENGTH = 8
+    const val MAX_PASSWORD_LENGTH = 20
+    const val PASSWORD_CRITERIA_COUNT = 3
+
+}
