+++ app/src/main/java/org/sopt/and/presentation/utils/Utils.kt
@@ -1,0 +1,8 @@
+package org.sopt.and.presentation.utils
+
+import android.content.Context
+import android.widget.Toast
+
+fun showToast(context: Context, message: String) {
+    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
+}
